+++ server/websocket/src/broadcast/mod.rs
@@ -1,6 +1,7 @@
 pub mod group;
 pub mod pool;
 mod publish;
+pub mod sub;
 pub mod tasks;
 pub mod types;
